Refactoring

+++ Decay.py
@@ -5,13 +5,13 @@ import math
 
 class Decay:
     @staticmethod
-    def linear_decay(t: int) -> float:
+    def linear_decay(t: float) -> float:
         return 1 / (t+1)
 
     @staticmethod
-    def exp_decay(t: int, time_const: float) -> float:
+    def exp_decay(t: float, time_const: float) -> float:
         return math.exp(-(t / time_const))
 
     @staticmethod
-    def power_series(t: int, epochs: int) -> float:
+    def power_series(t: float, epochs: int) -> float:
         return 0.005 ** (t / epochs)
